isolate main only network code
