+++ packages/insomnia/src/main/ipc/main.ts
@@ -1,9 +1,9 @@
 import { app, ipcMain } from 'electron';
 import { writeFile } from 'fs/promises';
 
-import { cancelCurlRequest, curlRequest } from '../../network/libcurl-promise';
 import { authorizeUserInWindow } from '../../network/o-auth-2/misc';
 import installPlugin from '../install-plugin';
+import { cancelCurlRequest, curlRequest } from '../network/libcurl-promise';
 
 export interface MainBridgeAPI {
   restart: () => void;
