+++ packages/insomnia/src/network/network.ts
@@ -28,39 +28,17 @@ import {
   RENDER_PURPOSE_NO_RENDER,
   RENDER_PURPOSE_SEND,
 } from '../common/render';
+import type { ResponsePatch, ResponseTimelineEntry } from '../main/network/libcurl-promise';
 import * as models from '../models';
 import { ClientCertificate } from '../models/client-certificate';
 import type { Environment } from '../models/environment';
 import type { Request } from '../models/request';
-import type { ResponseHeader } from '../models/response';
 import type { Settings } from '../models/settings';
 import { isWorkspace } from '../models/workspace';
 import * as pluginContexts from '../plugins/context/index';
 import * as plugins from '../plugins/index';
 import { getAuthHeader } from './authentication';
-import type { ResponseTimelineEntry } from './libcurl-promise';
 import { urlMatchesCertHost } from './url-matches-cert-host';
-
-export interface ResponsePatch {
-  bodyCompression?: 'zip' | null;
-  bodyPath?: string;
-  bytesContent?: number;
-  bytesRead?: number;
-  contentType?: string;
-  elapsedTime: number;
-  environmentId?: string | null;
-  error?: string;
-  headers?: ResponseHeader[];
-  httpVersion?: string;
-  message?: string;
-  parentId?: string;
-  settingSendCookies?: boolean;
-  settingStoreCookies?: boolean;
-  statusCode?: number;
-  statusMessage?: string;
-  timelinePath?: string;
-  url?: string;
-}
 
 // Time since user's last keypress to wait before making the request
 const MAX_DELAY_TIME = 1000;
@@ -96,7 +74,7 @@ export async function _actuallySend(
         // NOTE: conditionally use ipc bridge, renderer cannot import native modules directly
         const nodejsCancelCurlRequest = process.type === 'renderer'
           ? window.main.cancelCurlRequest
-          :  (await import('./libcurl-promise')).cancelCurlRequest;
+          :  (await import('../main/network/libcurl-promise')).cancelCurlRequest;
 
         nodejsCancelCurlRequest(renderedRequest._id);
         return resolve({
@@ -138,7 +116,7 @@ export async function _actuallySend(
       // NOTE: conditionally use ipc bridge, renderer cannot import native modules directly
       const nodejsCurlRequest = process.type === 'renderer'
         ? window.main.curlRequest
-        : (await import('./libcurl-promise')).curlRequest;
+        : (await import('../main/network/libcurl-promise')).curlRequest;
 
       const requestOptions = {
         requestId: renderedRequest._id,
